refactor: UI 관련 폴더명 및 폴더 위치 변경
